change: root page to app directory
